Split crc table generating code (MAKECRCH) out into separate tool makecrct.
